adk-agents/requirements.txt
requests
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
google-adk
google-genai>=1.49.0
google-cloud-firestore>=2.21.0